+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update `examples/one-page.html` to `examples/js/index.js` with new script helper.
 - Add support for Deck or Slide-level transitions
 - Add default transitions for Fade, Slide, and None
+- Fixes Full Screen component for Chrome/FF, adds support for Safari
 
 ## 6.0.0-alpha.7